+++ UI/src/mit-mainpanel/toolboxFoldout.tsx
@@ -4,8 +4,17 @@ import locale from "../lang/en-US.json";
 import { useLocalization } from "cs2/l10n";
 import { Foldout } from "./foldout/foldout";
 
+
+export const uilStandard =                          "coui://uil/Standard/";
+const toolSrc =                 uilStandard + "Tools.svg";
+const terrainHeightSrc = uilStandard + "NetworkGround.svg";
+const objectHeightSrc = uilStandard + "NoHeightLimit.svg";
+const rotateAtCenterSrc = uilStandard + "RotateAroundLeft.svg";
+const rotateInPlaceSrc = uilStandard +"ArrowCircularLeft.svg";
+
 export const ToolboxFoldout = (props: {state: FoldoutState}) : JSX.Element =>
 {
+
     // translation handling. Translates using locale keys that are defined in C# or fallback string from en-US.json.
     const { translate } = useLocalization();
 
@@ -16,6 +25,7 @@ export const ToolboxFoldout = (props: {state: FoldoutState}) : JSX.Element =>
         LabelLocaleKey: "MoveIt.TEXT_LABEL[Toolbox]", 
         LabelFallback: locale["MoveIt.TEXT_LABEL[Toolbox]"], 
         Tooltip: translate("MoveIt.TOOLTIP_DESCRIPTION[Toolbox]", locale["MoveIt.TOOLTIP_DESCRIPTION[Toolbox]"]),
+        Icon: toolSrc,
     }
 
     const TerrainHeight : FOMainEntryData = 
@@ -26,6 +36,7 @@ export const ToolboxFoldout = (props: {state: FoldoutState}) : JSX.Element =>
         LabelLocaleKey: "MoveIt.TEXT_LABEL[TerrainHeight]",
         LabelFallback: locale["MoveIt.TEXT_LABEL[TerrainHeight]"],
         Tooltip: translate("MoveIt.TOOLTIP_DESCRIPTION[TerrainHeight]", locale["MoveIt.TOOLTIP_DESCRIPTION[TerrainHeight]"]),
+        Icon: terrainHeightSrc,
     }
 
     const ObjectHeight : FOMainEntryData = 
@@ -36,16 +47,18 @@ export const ToolboxFoldout = (props: {state: FoldoutState}) : JSX.Element =>
         LabelLocaleKey: "MoveIt.TEXT_LABEL[ObjectHeight]",
         LabelFallback: locale["MoveIt.TEXT_LABEL[ObjectHeight]"],
         Tooltip: translate("MoveIt.TOOLTIP_DESCRIPTION[ObjectHeight]", locale["MoveIt.TOOLTIP_DESCRIPTION[ObjectHeight]"]),
+        Icon: objectHeightSrc,
     }
 
     const RotateAtCenter : FOMainEntryData = 
     {
         Section: "toolbox",
-        RawId: "rotateAtCentre",
+        RawId: "rotateAtCenter",
         Id: "rotateAtCentreRow",
         LabelLocaleKey: "MoveIt.TEXT_LABEL[RotateAtCenter]",
         LabelFallback: locale["MoveIt.TEXT_LABEL[RotateAtCenter]"],
         Tooltip: translate("MoveIt.TOOLTIP_DESCRIPTION[RotateAtCenter]", locale["MoveIt.TOOLTIP_DESCRIPTION[RotateAtCenter]"]),
+        Icon: rotateAtCenterSrc,
     }
 
     const RotateInPlace : FOMainEntryData = 
@@ -56,6 +69,7 @@ export const ToolboxFoldout = (props: {state: FoldoutState}) : JSX.Element =>
         LabelLocaleKey: "MoveIt.TEXT_LABEL[RotateInPlace]",
         LabelFallback: locale["MoveIt.TEXT_LABEL[RotateInPlace]"],
         Tooltip: translate("MoveIt.TOOLTIP_DESCRIPTION[RotateInPlace]", locale["MoveIt.TOOLTIP_DESCRIPTION[RotateInPlace]"]),
+        Icon: rotateInPlaceSrc,
     }
 
     const data : FoldoutData =