+++ Code/MoveIt/ChangeLog.md
@@ -1,2 +1,5 @@
 ﻿# Patch v0.5.7
-- Added Localization Support.
+* Added Localization Support.
+* Added Unified Icon Library Dependency
+* Added Icons to filters and toolbox actions.
+* Added Tooltip Icons and Toolbox tooltips.